--- models/database.py
@@ -1,3 +1,0 @@
-from flask.ext.sqlalchemy import SQLAlchemy
-
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:////tmp/test.db'
