+++ my_site/__init__.py
@@ -1,4 +1,1 @@
-from flask import Flask
 
-app = Flask(__name__)
-from app import views
