--- my_site/base.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Users table</title>
- </head>
-  <body>
-
-    <div class="container">
-
-      <!-- child template -->
-      {% block content %}{% endblock %}
-
-     
-    </div>
-
-  </body>
-</html>
