--- my_site/config.py
@@ -1,18 +1,0 @@
-# configurations
-"""
-class Config(object):
-
-
-	DEBUG = False
-	TESTING = False
-	DATABASE_URI = 'sqlite://:memory:'
-
-class ProductionConfig(Config):
-    DATABASE_URI = 'mysql://user@localhost/foo'
-
-class DevelopmentConfig(Config):
-    DEBUG = True
-
-class TestingConfig(Config):
-    TESTING = True
-"""
