--- my_site/user.py
@@ -1,14 +1,0 @@
-from main_app import db
-
-
-class User(db.Model):
-	id = db.Column(db.Integer, primary_key=True)
-	username = db.Column(db.String(80), unique=True)
-	email = db.Column(db.String(120), unique=True)
-
-	def __init__(self, username, email):
-		self.username = username
-		self.email = email
-
-	def __repr__(self):
-		return '<User %r>' % self.username
