--- my_site/user_form.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-  	<meta charset="utf-8">
-    <title>{{ title }}</title>
-  </head>
-  <body>
-  	<form action="{{ url_for('create_user') }}" method="POST" id="my-form">
-  		<input type="text" name="name" placeholder="Enter user's name">
-  		<input type="text" name="email" placeholder="Enter user's email">
-  		<input type="submit" value="Create user">
-  	</form>
-  </body>
-</html>
