--- my_site/users.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-  	<meta charset="utf-8">
-    <title>{{ title }}</title>
-  </head>
-  <body>
-  	  <a href="{{ url_for('create_user') }}">Create User</a>
-  	  <br>
-  	  <br>
-      <table border="1">
-      		<thead>
-      			<tr>
-      				<th>#</th>
-					<th>name</th>
-					<th>e-mail</th>
-				</tr>
-      		</thead>
-      		<tbody>
-      			{% for user in view_data['users']: %}
-      			<tr>
-      				<td>{{ loop.index }}.</td>
-					<td>{{ user.username }}</td>
-					<td>{{ user.email }}</td>
-				</tr>
-				{% endfor %}
-      		</tbody>
-			<tfoot>
-				<tr>
-					<td colspan="3">that's all</td>
-				</tr>
-			</tfoot>
-		</table>
-		<ul>
-			{% for friend_name in view_data['user_friends'] %}
-				<li>{{ loop.index}}. {{ friend_name }}</li>
-			{% endfor %}
-		</ul>
-  </body>
-</html>
